Incorporating Tore's updates to OMArc and GeoProj, making arc variables doubles to improve accuracy and appearance.
git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@1983 f68c23cc-7f87-4fbc-acfa-d9760b90acc1

+++ src/openmap/com/bbn/openmap/proj/GeoProj.java
@@ -1261,8 +1261,8 @@ public abstract class GeoProj extends Proj {
 
         switch (arcType) {
         case Arc2D.PIE:
-            rawllpts[rawllpts.length - 4] = (float) c.getRadLat();
-            rawllpts[rawllpts.length - 3] = (float) c.getRadLon();
+            rawllpts[rawllpts.length - 4] = c.getRadLat();
+            rawllpts[rawllpts.length - 3] = c.getRadLon();
         case Arc2D.CHORD:
             rawllpts[rawllpts.length - 2] = rawllpts[0];
             rawllpts[rawllpts.length - 1] = rawllpts[1];
